+++ lib/src/arithmetic/floating_point/floating_point.dart
@@ -7,3 +7,4 @@ export 'floating_point_adder_simple.dart';
 export 'floating_point_converter.dart';
 export 'floating_point_multiplier.dart';
 export 'floating_point_multiplier_simple.dart';
+export 'floating_point_rounding.dart';
